% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{bac_stre}
\alias{bac_stre}
\title{Streptococcus and Bacillus pangenomic data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81 rows and 149721 columns.}
\usage{
data(bac_stre)
}
\description{
A cluster membership data frame for genes from 81 bacterial genomes (rows) spread
in 149721 clusters (columns).
The dataset was created from publicly available data from Ensembl,It was produced
by a standard clustering pipeline with the default settings (BLAST and MCL).
and contains seven (81) strains of the following bacteria :
twelve (12) of Streptococcus pneumoniae, thirteen (13) of Streptococcus
Pyogenes, thirty nine (39) of Bacillus cereus and seventeen (17) of
Bacillus thuringiensis. Genome names are contained in the bac_stre_names
dataset.
}
\examples{
data(bac_stre)
}
\references{
Mpatziakas A, Psomopoulos FE, Moysiadis T and Sgardelis S. Computing pangenome statistics in R
 F1000Research 2017, 6(ISCB Comm J):1529 (poster) (doi: 10.7490/f1000research.1114765.1)
}
\keyword{datasets}
