% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{bac_stre_names}
\alias{bac_stre_names}
\title{Names of Streptococcus and Bacillus species}
\format{An object of class \code{character} of length 81.}
\usage{
data(bac_stre_names)
}
\description{
This dataset contains the names of the genomes inside the bac_stre dataset.
}
\examples{
data(bac_stre_names)
}
\references{
Mpatziakas A, Psomopoulos FE, Moysiadis T and Sgardelis S. Computing pangenome statistics in R
 F1000Research 2017, 6(ISCB Comm J):1529 (poster) (doi: 10.7490/f1000research.1114765.1)
}
\keyword{datasets}
