% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{binomixMachine}
\alias{binomixMachine}
\title{Binomix machine}
\usage{
binomixMachine(y, K, core.detect.prob = 1)
}
\description{
This function is a helper borrowed from package micropan to be used to compute
 binomial mixture pangenome models
}
\keyword{internal}
