% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_number}
\alias{cluster_number}
\title{Cluster Analysis with fluidity}
\usage{
cluster_number(fluidity_list, method = "ward.D")
}
\arguments{
\item{fluidity_list}{List produced by fluidity_all()}

\item{method}{Method of hierarchical clustering, see ?stats::hclust()}
}
\description{
This function outputs the proposed gene clustering based on fluidity distance
}
\examples{
\dontrun{cluster_number(fluidity_list)}

}
