% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grid_plot}
\alias{grid_plot}
\title{grid plot}
\usage{
grid_plot(Panmatrix, use_log)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{use_log}{Logical:Scale axis containing clusters with log}
}
\description{
This function outputs cluster tidyr::spread for genomes, genome participation per cluster,
gene participation per cluster and a brief summary allowing a quick exploration
of the available data
}
\examples{
\dontrun{grid_plot(panm)}
}
