% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gtools_comb}
\alias{gtools_comb}
\title{Helper function: gtools::combination}
\usage{
gtools_comb(n, r, v = 1:n, set = TRUE, repeats.allowed = FALSE)
}
\description{
This helper function is produces all possible combinations for a set of numbers.
  This is an direct copy of the combination() function from package gtools, and is
  used in function pm_fludity_all().
}
