% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_panmatrix}
\alias{make_panmatrix}
\title{Make panmatrix (MCL data)}
\usage{
make_panmatrix(x)
}
\arguments{
\item{file}{Disk Path to file}
}
\description{
This function allows importing MCL output
}
\examples{
make_panmatrix()

}
