% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_panmatrix_fami}
\alias{make_panmatrix_fami}
\title{Make panmatrix (fami MCL data)}
\usage{
make_panmatrix_fami(file)
}
\arguments{
\item{path}{file Path to file}
}
\description{
This function allows importing MCL output
}
\examples{
make_panmatrix_fami()

}
