% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_panmatrix_fami2}
\alias{make_panmatrix_fami2}
\title{Make panmatrix (fami 2 MCL data)}
\usage{
make_panmatrix_fami2(file_path)
}
\arguments{
\item{file_path}{file Path to file}
}
\description{
This function allows importing BLAST MCL output with default parameters
}
\examples{
\dontrun{make_panmatrix_fami2(file_path)}

}
