% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mg_plot}
\alias{mg_plot}
\title{Gene membership}
\usage{
mg_plot(Panmatrix, collapsed, use_log = TRUE)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{collapsed}{Defaults to True, sum the number of genes for all groups with more than
three times the number of the genomes explored}

\item{use_log}{Logical,scales axis containing clusters with log, defaults to FALSE}
}
\description{
This function produces a genes - membership plot  of the pangenome, i.e. the gene-cluster participation
frequency in the pangenome.
}
\examples{
\dontrun{mg_plot(Panmatrix)}
}
