% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{negTruncLogLike}
\alias{negTruncLogLike}
\title{Truncated log likelihood}
\usage{
negTruncLogLike(p, y, core.p)
}
\description{
This function is a helper borrowed from package micropan to be used to compute
 truncated log likelihood
}
\keyword{internal}
