% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{org_names}
\alias{org_names}
\title{Organism Names - MCL}
\usage{
org_names(file_path)
}
\arguments{
\item{file_path}{Disk Path to file}
}
\description{
This function outputs the genome names of a MCL clustering type input
}
\examples{
\dontrun{organism_names_panmatrix(file_path)}
}
