% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{organism_names_panmatrix_fami}
\alias{organism_names_panmatrix_fami}
\title{Organism Names - fami}
\usage{
organism_names_panmatrix_fami(file_path)
}
\arguments{
\item{file_path}{Path to file in disk}
}
\description{
This function outputs the genome names of a fami clustering type input
}
\examples{
\dontrun{organism_names_panmatrix_fami(file_path)}

}
