% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{panm_fluidity_all}
\alias{panm_fluidity_all}
\title{Genome Fluidity}
\usage{
panm_fluidity_all(panm)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{n.sim}{Number of simulations}
}
\description{
This function computes fluidity without sampling.
  Fluidity takes values in [0,1], with 1 denoting no common genes.
  This metric was introduced in:
  "Genomic fluidity: an integrative view of gene diversity within microbial populations" (Kislyuk et al ,2011)
}
\examples{
pm_fluidity(panm)

}
