% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_chao}
\alias{pm_chao}
\title{Chao lower bound estimator}
\usage{
pm_chao(panm)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function computes two versions of the lower bound Chao Estimator for the pangenome
   size along with variance and a 95% CI . This is a conservative estimator, ie. it is
   likely to produce a small estimation of the pangenome compared to other estimators.
   This function is an optimized and expanded version of chao() from package micropan.
}
\examples{
pm_chao(panm)

}
