% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_heaps}
\alias{pm_heaps}
\title{Heap's Law fitting}
\usage{
pm_heaps(panmatrix, n_perm)
}
\arguments{
\item{n_perm}{Number of permutations}

\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function determines the closedness of the pangenome using a curve fit of the data
   based on Heaps Law.
}
\details{
The regression fit returns two estimated parameters, intercept and decay parameter a.  If a<1 then the pangenome is considered to be open. This is an optimized version
   of the heaps() function from package micropan. The theoretical aspects are discussed in the canonical work of Tettelin et  al. (2008)
}
\examples{
pm_heaps(panm,100)

}
\references{
Tettelin, H., Riley, D., Cattuto, C., Medini, D. (2008). Comparative genomics: the bacterial pan-genome. Current Opinions in Microbiology, 12:472-477.
}
