% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_plot}
\alias{pm_plot}
\title{Panmatrix Plots}
\usage{
pm_plot(object, show_cluster, plot_type)
}
\arguments{
\item{show_cluster}{Integer value  optional parameter allows to user to "zoom",
ignoring clusters that have organism participation below it}

\item{plot_type}{Can be either line or bar}

\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function produces a basic plot of the pangenome.
   This can be either a line plot or a bar plot.
}
\examples{
make_panmatrix_fami()

}
