% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_panmatrix}
\alias{make_panmatrix}
\title{Make panmatrix (MCL data)}
\usage{
make_panmatrix(x)
}
\arguments{
\item{file}{Disk Path to file}
}
\description{
This function allows importing MCL output
}
\note{
MCL output as described in F. E. Psomopoulos, O. T. Vrousgou, and P. A. Mitkas, "Large-scale modular comparative genomics: the Grid approach [v1; not peer reviewed]," F1000research 2015, vol. 4(ISCB Com, iss. 377, p. 1, 2015. doi:10.7490/f1000research.1110127.1
       A. M. Kintsakis, F. E. Psomopoulos, and P. A. Mitkas, "Data-aware optimization of bioinformatics workflows in hybrid clouds," Journal of big data, vol. 3, iss. 20, pp. 1-26, 2016. doi:10.3389/fpls.2016.00554
}
\examples{
make_panmatrix()
}
