% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_panmatrix_fami}
\alias{make_panmatrix_fami}
\title{Make panmatrix (fami MCL data)}
\usage{
make_panmatrix_fami(file_path)
}
\arguments{
\item{file_path}{Disk path to file}
}
\description{
This function allows importing MCL output
}
\examples{
\dontrun{make_panmatrix_fami(file_path)}

}
