% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mg_plot}
\alias{mg_plot}
\title{Gene membership}
\usage{
mg_plot(object, collapsed, use_log)
}
\arguments{
\item{use_log}{Logical:Scale axis containing clusters with log, defaults to FALSE}

\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{Collapsed}{Defaults to True, sum the number of genes for all groups with more than
three times the number of the genomes explored}
}
\description{
This function produces a genes - membership plot  of the pangenome, i.e. the gene-cluster participation
frequency in the pangenome.
}
\examples{
mg_plot(panm)
}
