% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{org_names}
\alias{org_names}
\title{Organism Names - MCL}
\usage{
org_names(x)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{n.sim}{Number of simulations}
}
\description{
This function outputs the genome names of a MVL clustering type input
}
\examples{
organism_names_panmatrix(file)
}
