% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{org_names_fami2}
\alias{org_names_fami2}
\title{Organism Names - MCL (fami2)}
\usage{
org_names_fami2(file_path)
}
\arguments{
\item{file_path}{Disk Path to file}
}
\description{
This function outputs the genome names of a MCL clustering type input
}
\examples{
\dontrun{organism_names_fami2(file_path)}
}
