% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{organism_names_panmatrix_fami}
\alias{organism_names_panmatrix_fami}
\title{Organism Names - fami}
\usage{
organism_names_panmatrix_fami(file)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{n.sim}{Number of simulations}
}
\description{
This function outputs the genome names of a fami clustering type input
}
\examples{
organism_names_panmatrix_fami(file)

}
