% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{panm_summary}
\alias{panm_summary}
\title{Panmatrix Summary}
\usage{
panm_summary(object, ...)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function produces a frequency table of Genome participation in clusters
}
\examples{
make_panmatrix_fami()

}
