% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_binom}
\alias{pm_binom}
\title{Binomial mixture fitting}
\usage{
pm_binom(Panmatrix, K.range = 2:8, core.detect.prob = 1)
}
\arguments{
\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}

\item{K.range}{Range of model components to be tested, defaults to search from 2 to 8}

\item{core.detect.prob}{Detection probability of core genes. Defaults to 1.0}
}
\description{
This function determines the closedness of the pangenome using a binomial mixture fit of the data.
}
\details{
The function returns pangenome and core size estimation along with the propabilities
 of the components.This is an optimized version of the binomixEstimate() function from package
 micropan  (Snipen & Hiland, 2015).
}
\examples{
\dontrun{pm_binom(panm, K.range =2:8)}

}
\references{
L. Snipen and K. H. Liland, "micropan: an R-package for microbial pan-genomics.," BMC bioinformatics, vol. 16, p. 79, 2015
}
