% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_binom}
\alias{pm_binom}
\title{Binomial mixture fitting}
\usage{
pm_binom(pan.matrix, K.range = 3:5, core.detect.prob = 1, verbose = TRUE)
}
\arguments{
\item{K.range}{Range of model components to be tested}

\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function determines the closedness of the pangenome using a binomial mixture fit of the data.
}
\details{
The function returns pangenome and core size estimation along with the propabilities
 of the components.This is an optimized version of the binomixEstimate() function from package
 micropan  (Snipen & Hiland, 2015).
}
\examples{
pm_binom(panm, K.range =2:8)

}
\references{
L. Snipen and K. H. Liland, "micropan: an R-package for microbial pan-genomics.," BMC bioinformatics, vol. 16, p. 79, 2015
}
