% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pm_fluidity}
\alias{pm_fluidity}
\title{Genome Fluidity with sampling}
\usage{
pm_fluidity(panm, n.sim = 10)
}
\arguments{
\item{n.sim}{Number of simulations}

\item{Panmatrix}{Panmatrix produced by make_panmatrix functions}
}
\description{
This function computes fluidity with sampling, as implemented on package micropan (optimized for speed).
  Fluidity takes values in [0,1], with 1 denoting no common genes.
}
\examples{
pm_fluidity(panm, n.sim=100)

}
\references{
A. O. Kislyuk, B. Haegeman, N. H. Bergman, and J. S. Weitz, "Genomic fluidity???: an integrative view of gene diversity within microbial populations," BMC genomics, pp. 12-32, 2011.
M. M. Deza and E. Deza, Encyclopedia of Distances. Springer, 2009.
}
